VK Backend ML/requirements.txt
vk_api
python-dotenv
SQLAlchemy
psycopg2-binary
requests